[FIX] web: initial rpcs are done twice on /web/webclient/qweb/
The purpose of this commit is to avoid double calling
on `/web/webclient/qweb/` during initialization (one in legacy and one
in the new code) and to load the qweb templates as quickly as possible.

To do this, we will take inspiration from `odoo.loadMenusPromis`
by creating a promise in the `web.webclient_bootstrap` template.