[FIX] payment_stripe: load Stripe.js library on the payment form
When the payment form is displayed, Stripe.js library isn't loaded, thus preventing the customer to pay online.

When the refactor of assets was done, the link to the Stripe library was moved to __manifest__.py.
As the link does not end with '.js', the file wasn't loaded anymore.

See PR #60632

X-original-commit: 8535b090a36fe6472c4b3d8211f3be39c7f1c58a