[IMP] web: Add lint/format tooling for JS
Add features to improve the dx of js devs.
It is all stored in web under the tooling dir. It has three bash scripts:
- enable: move needed files to the root dir to activate the tooling env
- disable: remove all the files from the root dir
- reload: disable then enable to get the latest changes

Features:
- ESLint and Prettier config
The idea is to use the same configs for all the devs.
It is possible to have sub config files if a module wants it's own
format rules.
Some configuration on the dev editor is most likely needed for all that
to work (plugin eslint, changes in .vscode/settings.json, etc.)
Example of a  .vscode/settings.json (located higher than community and enterprise):
{
  "editor.defaultFormatter": "dbaeumer.vscode-eslint",
  "editor.formatOnSave": false,
  "eslint.workingDirectories": [
    "community", "enterprise",
  ]
}

- Script to see if web is all well formated
- Script to format all web
- Script to format all staged files

- Pre-commit hook that execute script to format all staged files
- Pre-push hook that ask confirmation on master-wowl force push
Both hooks are in .husky and not in .git/hooks, meaning only people
using the npm install will have them active.

Notes:
- This impacts only community, and should be replicated in web_enterprise.