[FIX] web: actions/views legacy layer
OWL and legacy interface is tricky: on the one hand we need to support that the state of a legacy controller
is encapsualted within the controller widget itself (in particular the legacy Model).
On the other hand, we need to accomodate to owl's dom management.
We thus need to transmit alive legacy widget (and potentially with their element still in DOM) across
new OWL Components.

This leads to some heavy plumbing in the different layers, which mostly relies on
the t-directive's and VDOM's implementation details.
In particular that High Order Component's elements are transmitted up to the first concrete Component
through their "dummy" vnode.

Other implementation detail's involve the destruction of the components in some circumstances, essentially triggered
by competing renderings.

This commit is not aimed to be understood by everyone :D