[IMP] crm: Reducing the amount of emails sent for the task assignement by introducing a boolean.
PURPOSE

It turns out that the user can quickly get spammed with "You have been
assigned to" emails. The problem is that the user has no proper way to
avoid sending them. As many tasks can be assigned by day, these emails
can consume a large part of the emails quota that each Odoo instance has.

With these new changes, a new boolean will be added on the wizard that
allow the user to schedule a new activity. By unchecking that boolean,
the user can ask the server to not send a notification by email to the
user that gets assigned to the activity. The user will be able to reduce
the spam and limit the amount of emails sent for the task assignment.

SPECS

- Add a boolean field on the `mail.activity` model that will prevent
  the server to send a notification by email to the user that gets
  assigned to a new activity.

LINKS

Task id: 2209392