[FIX] web: sale_product_configurator_edition_tour
In the action manager, actions that were stored in the actionCache were not immutable.
The `sale_product_configurator_edition_tour` brought light on this due to the fact that the product configurator action was keeping its `context.configuratorMode` to the `"add"` value instead of `"edit"` for subsequent loadActions.

Furthermore, the actionCache key generation was missing the additional_context information.