[IMP] sale_timesheet{,_purchase,_expense}: extend report
Prior to this commit, the project profitability report view was not
extensible.

With this commit, it becomes possible to add jointures to the view and
fields to the select statement.
With this commit, the expenses related to the analytic account which
does have the same product_id, the same quantity can be used to group
the lines.
With this commit, the purchase orders which are related to a
sale_order_line can be used to group the lines.
With this commit, the vendor bills which are related to a
sale_order_line can be used to group this lines.

task-2458135 - part:8