POC héritage with decorator
The goal of this commit is to implement an inheritance system between the edi_formats.
For now, it only concerns _get_xml_builder : when the function is called on an edi_format, it will automatically call the parents in order first, then pass the result to the child so it can modify it.
The motivation behind this is : if we only do api.model function, for example when we call _get_nlcius_builder(), if we want to modify the party after the generation, we have to go through each party and edit it's value, most
likely by calling a function that will do this. This adds a lot of functions and makes the codes somewhat hard to read. It would be nice to have only one function that is overriden. That way, _get_nlcius_builder wouldn't even have
to call the specific function since everything would be handled by the generic functions.
Same goes for the xml_generation => we always need a function that will call _get_ubl_2_0_builder() then build. It would be nice that we just have to call _get_xml_builder and based on self.code, the correct xml is generated
(this is what this POC does).

Ideas :
- do the same for other function. For now, we only do it for _get_xml_builder, but we could imagine inspect the name of the function as well and call the right function (to avoid creating multiple decorators).
	- OR (better idea !!) call the function like _get_xml_builder_%s % parent_code
- we could get rid of the if self.code != 'nlcius': return super() and stuff like that for ALL the methods in account_edi_format.
- maybe we could get rid of the decorator and its parameter (it's boring that we have to type the code in the decorator parameter) => make our custom super() function that when called :
	parent = super(edi_format) would return the correct class where the methods are defined. So that we would just need to do
	parent.my_method(self, *args) => remember that parent is a class not an object. All objects have the type odoo.api.account.edi.format so it doesn't work, we need for example <class 'odoo.addons.account_edi_ubl.models.account_edi_format.AccountEdiFormat'>
I think a parameter for the decorator would still be needed, but it could be on the class instead. We need somehow to say that this method belongs to this format. Another solution would be to put it directly in the name of the
class and do some introspection based on that but it might not be a good idea.
- Create a mechanism so that super is not called (would work if we do the last point) => par exemple, si on doit tout recalculer pour un truc (par exemple les taxes), on n'appelle pas super. Pour l'instant super est appelé et on
peut overrider complètement mais alors c'est dommage d'avoir fait les calculs de super.