[FIX] project: fix test project update ui
Before this commit, the test_project_update_ui sometimes crashes because
the render of the right side panel is not done before the step using
this.

This commit adds a timeout to wait the render of the right side panel.
With this timeout, all the steps in the project_update_tour should
correctly be done.

Related PR: #68899