[IMP] web: make startServices promise reject when missing dependencies
Previously, the promise returned by startServices was resolved when some
services were missing dependencies. The rationale was that those
dependencies could be fulfilled by bundles that might be in the page
later.

This rationale ignored the fact that bundles or scripts that are in the
same page will be executed in the same microtick.

This commit changes the returned promise from startServices such that it
will throw if after all services that were registered in the registry at
call time that could be started are started, and there are still
services registered that could not be started for lack of dependencies.