[IMP] web: automatically add services dependencies in test env
Before this commit, it worked only for the first level, i.e. if
there was a service A in the registry, which depends on a service
B, B was added to the registry. But if B had a dependency on C, C
wasn't added to the registry.

closes odoo-dev/odoo#919

Related: odoo-dev/enterprise#145
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>