[FIX] web_editor: fix convert_inline on Safari iOS
Any CSS rule that is not understood by Webkit will usually be
ignored by the renderer, but that is not the case when the
selector is explicitely matched against in pure javascript. In
such cases, the engine throws a SyntaxError, resulting in a
traceback. The code above will match against such selector as
long as they are used in the CSS of Odoo alongside the editor.
This try catch protects against getting a traceback in those
cases on Safari and other webkit-based browsers.

Steps to reproduce the issue on Safari iOS:
- Open a task in Project
- Click Send a Message
- Click on the enlarge button to use the full composer
- Traceback