[FIX] mail: translate mail template based on recipient language
PURPOSE

It turns out that some parts of the notification emails (i.e: the label of the
buttons, the access links and the signature) are not translated in the language
of the users who will receive the emails. The system is currently using the
language of the user who wrote the message.

The system will now group the recipients by language and translate the template
in the language of each group. The buttons, the access links and the signature
will also be translated.

SPECS

- Translate the label of the buttons, the access links and the signature of the
  emails in the language of the recipients.

LINKS

Task id: 2555155