[FIX] auth_ldap: catch all ldap errors
When making an authentication using LDAP, python-ldap raises
exceptions in case of failure (wrong password, bad configuration,...).
The `_authenticate` methods returns False in case of invalid
credentials but does not for other errors.

While not exploitable in practice, it may be possible for a server to
fail (e.g. connexion error,...) between the `_get_entry` and
`_connect` call, returning the result of `_get_entry` instead of
failing.
Remove the log that was unnecessary (as the possible errors would
already be caught during the `_query(conf, filter)` call.

Introduced at 8bff93feee0c40dd9

Sign the Odoo CLA