-- COOKIE OF BSO --
[IMP] base, web, website, *: differentiate essential & optional cookies

*: im_livechat, survey, utm, website_crm_iap_reveal, website_forum,
   website_livechat, website_sale, website_sale_comparison

Before this commit all cookies when considered essential.

This commit makes some of them optional. It also adds the possibility
for activating a second type of cookie bar that makes it possible for
the website visitor to only accept the essential cookies.

task-2800976

Co-authored-by: kzh <kzh@odoo.com>
Co-authored-by: bso <bso@odoo.com>

[FIX] website: prevent Google Analytics from kicking in before consent

This commit prevents Google Analytics from being enabled if optional
cookies are declined.

task-2800976

[REF] web, *: remove cookie utils from general purpose utils

*: im_livechat, survey, website, website_forum, website_livechat,
   website_sale, website_sale_comparison

Since [1] the cookie methods could be accessed by either importing them
from `cookie_utils` or importing them from `utils`.
This makes it confusing when one of the methods needs to be patched.

This commit removes the cookie utils from utils and adapts all calls
accordingly.

[1]: https://github.com/odoo/odoo/commit/f14bbf1bdc70873ddd8915516c5c3ab6ae72a796

task-2800976