[FIX]stock: fix inconsistency beteen product variant and product template.
Before commit:
The issue is that the product is not in the 'available' filter but its kanban tile says 'on hand' quantity is available. A different issue is a raise because of this commit: https://github.com/odoo/odoo/commit/b08e3d114aa4efe687b4d4cca3de462a3dbe6eea. in this commit set the
"active_test=False" for product template that's why it fetches all active and inactive warehouse and it becomes true for product variant as we don't get active_test=False. so condition become true and it will fetch only active warehouse and related location and result is
displayed the different on_hand qty on variant and template.

After commit:
we set the 'active_test=False' on variant and 'action_open_quants' also to keep the consistency on variant and template on_hand quantity and it's related quants too.