[FIX] web: image with new one at the same url
Currently, when the user updates the task to add an image, the new image not
going to visible in the kanban view once image is uploaded without forcing the
user to reload the page

This is happening because of the same URL for the images. Due to cache, it will
display the old one instead newly uploaded one.

With this commit, put a unique URL and now a new image will
going to be visible in the kanban view once an image is uploaded without forcing
the user to reload the page