[IMP] website_sale: filter products by price with custom range input
When browsing the eCommerce part of the website, customers should be able to filter shop products based on base sale price. The filter slider can be enabled through the customize menu and it only filters using the product's base price (for performance reason) and takes into account currency conversions, not pricelists and discounts. It uses a custom-built library for the range input with 2 cursors (min. price and max. price).

The slider is hidden when all products shown on the current page (category, etc.) have the same price (or there is only one product).

A more advanced version of this implementation would enable the filtering to take more factors into account.