[REF] web: introduce ControlPanelBottomContent
The ControlPanel now supports being passed an object "bottom-content" in its display props.
That Object is constituted by a component and by props to pass to it.

The bottom content of the Control panel is rendered as a Slot within that component

If not present, the ControlPanel continues to display its bottom part as usual