[IMP] auth_totp: allow remembering trusted devices
- Added the 'Trusted Devices' feature
- Added 'Remember this Device' checkbox on /web/login/totp
- Added trusted device OS / browser on Profile > Account Security

This feature allows users to remember their device to bypass the 2FA
for the next connections. The trusted devices are displayed in a
'Trusted Devices' One2Many under the 'Developer API Keys'.
It is possible to revoke all the trusted devices at once with a
special button. It is also possible to revoke one at a time on the
desired one.