[FIX] survey: fix the timer in survey
Before this commit, when someone from audience manually manipulates
the system time. The current timer calculation actually depends
on both the server time, and the system time of audience.

In this commit we fix issue by taking the offset of server utc time
and user's system utc time and adding this offset in countDownDate.
If user does not restart his server after merging this code then there
will be '404 Not Found' error as our partial code is in python which,
will be not updated. We handle this fallback here through try and catch.

Task Id: 2612972