[IMP] web: Add action and view to search model
This commit adds the additional `action` and `view` properties to the
search model. Both of them are expected to be objects with an `id` and
a `type` property (`action.id` replaces the previous `actionId`).

These objects allow components using the search model to have access to
the action/view's type and id.

The default value for both of these keys is an object with `id: false`
for ease of use.

Part-of: odoo-dev/odoo#1024