[FIX] website: fix conversion of specific svgs on firefox
A function allowing to convert a SVG to a PNG has been added as
an utility function in website.utils and is used for the upload
of document in website_slides. This function add support for a
specific case which was not supported by the previous conversion
function used for the document upload in website_slides: SVG
without width and height attributes on Firefox.

task-2602521