[IMP] portal, *: overview revamp
`*`: account, auth_totp_portal, hr_timesheet, payment, project, purchase, sale,
sale_timesheet, website_crm_partner_assign, website_payment, website_sale

The page /my /my/home is retitled Overview instead of Documents.

This overview is presented as a kanban-view menu (with card-like elements)

Each card links to a page of the portal and contains preview information about
the content of the linked page in the form of counters and descriptions

Setup for a new card in the overview menu:
  - The card expects a `description_data` which should be a list of dictionaries
    containing 2 keys: description (mandatory) and counter (not mandatory).
  - The value for this description_data should be defined in
    _prepare_portal_overview_values and set in the template
    portal.portal_docs_entry
  - The list represents each line of the description of the card.
  - Since counters are computed asynchronously to improve performance, the
    counter should be the name of the counter that will later be computed in
    _prepare_portal_counter_values

The last card "My Account" links to /my/account which now contains
the information that was previously in the side-bar of the old "Documents"
page.

The card "Leads" and all its related routes are now disabled when the option is
not activated in the settings in backend.

The side bar is reworked to contain a small uniformized menu with the following
links:
  - My Address
  - Payment methods
  - Password & Security

portal.portal_docs_entry is removed. Instead, a new view
portal.portal_overview_card is added for the new kanban-like menu.

Normalize the breadcrumbs for those pages (the home link and the page title are
displayed).

portal.portal_breadcrumbs is renamed portal_breadcrumb. To create a page with
breadcrumbs, portal.portal_layout should be used and portal.portal_breadcrumb
should be overrided to set the desired breadcrumbs behavior.

The breadcrumbs for these pages are updated to this new convention:
  - /donation/pay
  - /my/payment_methods
  - /payment/confirmation
  - /payment/pay
  - /payment/status

task-2490142