[IMP] project: new ProjectControlPanel class
The reporting views have been converted to OWL so that their extensions
like the project_graph and project_pivot views have to be converted too.
We thus need an owl version of the ProjectControlPanel class.

Here we introduce that control panel and refactor a bit the legacy code.

Part-of: odoo-dev/odoo#1022