[IMP] crm: redefine won/lost and apply rules in PLS frequencies/views
Purpose
=======

Redefine Won / Lost heuristic, apply it everywhere in computations, and align
the display (buttons and ribbons) with it.

Specification
=============

New heuristic:
- Lost: Proba = 0 AND archived (previsously: archived)
- Won: won stage AND 100% (was previously: won stage and active) (*)

That means that even if a lead is archived, it is still considered as won
if it is in won stage. That also mean that we should block the edition of
probability in won stage as it should always be 100%. Won / lost ribbons
display rules should be aligned to the new heuristics.

Update on rules for won / lost frequencies:
We handle won and lost frequencies in the create and write methods. Based
on the old / new values and new heuristic, we check all the followings
and update the frequency table each time a lead leaves / reaches a won or
lost state. Several can happen simultaneously.

- Set to Won Stage: Reach Won (and P will be 100)
- Move out from Won stage: Leave Won
- Archive: If P is 0, Reach Lost
- Unarchive: If P was 0, Leave Lost

Clarify the Restore action (vs unarchive):
- Unarchive <=> Set active = True
- Restore <=> Unarchive + Force automated probability (and its recomputation).
Restore button is displayed only on lost leads.

This commit adds two computed fields on crm.lead to ease domains to
display/hide fields, buttons and ribbons, and to ease the reading
of domains based on leads being won / lost. Both have a search method to
allow to search on them and to use them in those domains:
- is_lost (Tracking enbaled to follow the state of the lead in the chatter.)
- is_won

The new heuristic also applied in other places:
- Ribbons on the lead views
- Filters 'Won' / 'Lost' / 'Ongoing' on the crm.lead model
    - Also on the filters for portal users, on their portal.
- Domains including / not including lost / won leads.

Constraint (*)
=============

In order to have a perfect equivalence between 'won' lead and won stage,
we now add a constraint to force 100% probability for leads in a won
stage. An upgrade script is added to ensure this as well. We can therefore
consider that a won stage implies 100% probability

Why oh why ?
============

Before this change, the old won/lost rules where the following:
Lost: Proba = 0 OR archived
Won: won stage AND active

Which means that both rules could have coexisted. If the lead were in won stage,
active but probability = 0%. (that should not happens but was still allowed).
The objective with the new rules is to use the probability as a necessary
but not sufficient criteriat. Probability makes it impossible to be both
lost and won. But it is not sufficient to classify a lead as won or lost:
- 0% does not mean it's lost, it must also be archived.
- 100% does not mean it's won, it must also be in won stage.

That will avoid having automatically lost or won leads. If Predictive lead scoring
computes 0% or 100% chance of winning, the lead will not be flagged yet as lost or
won and the user will still need to take action to correctly classify the lead.

Tests
=====

Tests are updated to match the new heuristic.
Some tests are clarified and comments added to ease further work
New tests are added to check new fields and assert new heuristic.

Task-2654916
COM PR odoo/odoo#77089
UPG PR odoo/upgrade#4074