[FIX] industry_fsm: fix locale for start hour displayed on the kanban card
Description of the issue/feature this PR addresses:
In the kanban view of industry_fsm, the start time is hardcoded in 12 hours format.
The start time should be based on the user's time format

Current behavior before PR:
The start time is displayed in 12 hours format

Desired behavior after PR is merged:
The start time is displayed in 12 or 24 hours format according to the user's preferences.
The date is displayed according to user's preferences

task-2634340