[FIX] stock: wrong domain for related company_id when check company
On stock.storage.category, check_company is true for package_type_id and
product_id, a domain is automatically created for them to restrict the
company. However, the company_id on stock.storage.category is a related
field, when create new record, it's always False. As a result, only
package type and product without company_id can be selected. In this
commit, we set default_company_id in the view to avoid this issue.

Task-2614519
PR #76645