[FIX][mrp]: Fix issue for serial number in MO
Before this commit, if another product which is not tracked by serial no. in that case it's lot_id is FALSE and if we search from
stock.move.line then we can search move line which is having FALSE lot_id.
And, while we create the lot using specifying the lot name from MO at that time move line doesn't have the lot_id.
so we can not create the lot_id.

For example: product A not tracked by serial number and product B tracked by serial number.
and move line is available for product A with the done quantity 1 and state is 'done' and lot_id FALSE.
in this case while we create a lot id for product B from MO then we will get warning like
```
The serial number False used for component B has already been consumed
```

After this commit, search only that move lines for products which is tracked by serial number.

opw- 2614103