[FIX] base: preserve Outlook conditional comments on send mass mailing
For email design in the context of Microsoft Outlook, we want to keep
some magic Microsoft comments (Outlook conditional comment), which -
until this commit - were skipped by QWeb. These allow us to change the
rendering exclusively for Outlook so as to overcome some of its
limitations. This commits introduces a qweb rendering option
(`preserve_comments`) for when - like in mass mailing - we want to keep
comments.