[FIX] web_editor: dotted line persistent even after link tools destroy
Currently, when you add a link to any of the text and click outside of
the linktools area, it will destroy the link tools and add the link to
the target. But the dotted line (.oe_edited_link) keep as it is even after
link tools destroy.

For website, we have cleanForSave method to clean up the .oe_edited_link
before going to save the template. but for Mass Mailing, there is no any
provision to cleanup. so on destroy we are just removing the .oe_edited_link
after the set Focus method.