[FIX] web: run server actions in the correct user context
Currently, actions are run only with their own specified context (aka ctxt in the context field).
This means that the user context is not provided (including lang, tz, and current companies).

This means that if any custom logic of the server actions depends on the user context, it won't work
properly.

This commit makes sure that the user context is correctly conveyed to the server when running server actions,
making sure the language, tz, and companies in the context (and environment) are correct.