-- DYNAMIC SNIPPET OF ARD --
[IMP] website, *: give flexibility to dynamic snippet templates

*: website_sale,website_blog,website_event

Commit [1] introduced new templates. This commit's goal is to build on
that template system to provide more flexibility for designers.
It adds new data-attributes to set on the first node of a template.

- data-row-per-slide: Allows for the template to tell a dynamic snippet
carousel how many rows of cards it should display.

- data-arrow-position: Allows for the template to tell a dynamic snippet
carousel where the arrows should be positioned.
(for now just bottom, or side)

- data-extra-classes: give the template the possibility to add extra
classes to the dynamic snippet instead of just the card.

This commit also removes options to select the amount of elements
displayed on each row. This is now only changeable through templates, or
by directly editing the snippet's DOM through debug tools. This was done
to reduce the amount of options visible, so that the user just chooses
a template and it is ready to go.

[1]: https://github.com/odoo/odoo/commit/49bd79b0bdca76415780bdfa199195371c2692e5

task-2654924

[IMP] website_sale: make cta button work on debug dynamic snippet

Previously, on the debug dynamic snippet, you could choose a product
filter and a template that had an add2cart button but that button
would not work.

This commit fixes that by creating a widget per card and moving the
card specific code to this widget.

task-2654924

[IMP] website_sale: improve design of templates

This commit improves the design of dynamic products templates.
It introduces templates using multiples rows and fine tunes the
design of the other ones.
To make thumbnails work in the template selector, an extra
data-attribute can now be processed by the snippet option.

- data-thumb: it references the location of the thumbnail for a
template. If present the option will show the svg / img.
If not it will display the name of the template.

task-2677203

Co-authored-by: Arthur Detroux <ard@odoo.com>

[IMP] website_sale: add tour for snippet products

This PR added new templates and features to the dynamic snippet
products. The goal of this commit is to test these features by selecting
every newly made template, saving, and adding a product to the cart.

Another tour was added to test the functionality of the recently
viewed product filter.

task-2677203

[REF] website: review options for the dynamic snippet

Prior to this commit, the Dynamic Snippet options would modify the DOM
at places were it's not supposed to. This commit aims at making the
option follow the flow of an option:

- Modify the target's DOM either on build or on a change of option
- Modify the UI either during willStart or updateUI

To do so, the methods that fetch the template data needed to be split
form the methods that create the buttons.

The benefits are that the option is less likely to cause race condition
and errors while performing its task.

task-2677203

[FIX] website: make dynamic snippet carousel slide again

After commit [1] upgrade Bootstrap to 5.1, the attributes used for
Bootstrap Components changed [2].

This commit fixes those attributes and also make the snippet
automatically slide on page load as per the specs on the task.

[1]: https://github.com/odoo/odoo/commit/1fcd098af5fa18e848c66bb83bdceb6ba0308512
[2]: https://getbootstrap.com/docs/5.1/migration/#javascript

task-2677203

[FIX] website, website_payment: correctly add frontend template

Commit [1] changed the way QWeb templates are loaded for public widgets,
removing the xmlDependencies attribute.

Most of the public widgets' templates were therefore defined in
assets_wysiwyg. This is wrong however, as those assets are only loaded
in specific situations.

This commit removes the templates definitions from the assets_wysiwyg
into specific records defined alongside the snippet templates.

Steps to reproduce:
- Drop the "Image Wall" snippet
- Save
- Visit as a public user (logged out)
- Click on an image from the image wall
- TB

[1]: https://github.com/odoo/odoo/commit/f05adbc8b59c67cbec4847687a1ee07e620c2a7a

task-2984724