[FIX] website: translate configurator's features page
The configurator was introduced at [1] and with it, added a page that
displays the features and pages that can be added to a website through
said configurator.
This page wasn't translated.

Steps to reproduce:
    - Have a user with main language not being english
    - Create a new website with said user
    - Go through the configurator
    - "Pages and Features" page is not translated properly

Calls to the model were not using the right environment
(specifically at website/models/website.py:L267[2]).
This commit fixes that by using the right env to make calls within the
configurator code.

[1]: e8a5af2e28a4724f86783746706cb59175f086bf
[2]: https://github.com/odoo/odoo/commit/e8a5af2e28a4724f86783746706cb59175f086bf#diff-5e92e473fa4d3da6db7ef727fb217dad51ef6c2383913edca73fe040a23e82c2R267

task-2687416