[POC] im_livechat: make 'questions chaining' work
This commit is a PoC and an attempt to solve an issue we have with survey and
that we are about to have in the chatbot as well.

When creating questions for the chatbot, you can set a "triggering answer",
meaning for example that you want to show the "Please explain your issue"
question only if the user has answered "I have an issue" to the previous
question.

In order to do that, you have to manually save the form every time a new
question is created, so that, in your next question, you will be able to select
one of the answers of the previous question for your "triggering answer".

That solves half the problem.

The other half is that you want to add a domain to only show answers to
question *of that particular chatbot*.
Indeed, it would be pointless to show answers to questions of other chatbots
that operate in completely different scripts.

To solve that, we need to manually fetch the freshly created chatbot_id from
the model and add a manual context key that is verified in a _name_search
override of the question.answer model.

Setting a regular domain on the triggering_answer_id field does NOT work, as
the web client will pass 'chatbot_id', '=', false.

Task-2030386