[FIX] base: do not create modules in noupdate
The purpose of noupdate is to allow users to customize the value and
ensure it won't be rollbacked to it's previous in case of update.
Having the modules in the noupdate is not useful as not something a
user will change.

The drawback of noupdate is that users with an existing db won't
benefit from changes made in the manifest (e.g. description of the
module) or get new translations.