[IMP] mrp: Remove Mass Produce limitations
The initial release of Mass Produce has limitations :
- no serial-tracked components,
- no more than 1 lot for lot-tracked components

This commit removes those contraints.

Additionally, the code has been refactored to use batch creation.
The performance gain depends on the complexity of the bom / reservation and the quantity to be produced.
Compared with v1, tests show times have been divided by 5 to 50+ depending on quantity...

task 2633369